补充增加遗漏的 intptr_t 转换，并且考虑到不同 MPI 的实现，现在都统一增加 intptr_t 的转换

+++ src/main/resources/assets/cpointer/src/jse_clib_CPointer.c
@@ -6,10 +6,10 @@ extern "C" {
 #endif
 
 JNIEXPORT jlong JNICALL Java_jse_clib_CPointer_malloc_1(JNIEnv *aEnv, jclass aClazz, jint aCount) {
-    return (jlong)MALLOCN(aCount, 1);
+    return (jlong)(intptr_t)MALLOCN(aCount, 1);
 }
 JNIEXPORT jlong JNICALL Java_jse_clib_CPointer_calloc_1(JNIEnv *aEnv, jclass aClazz, jint aCount) {
-    return (jlong)CALLOC(aCount, 1);
+    return (jlong)(intptr_t)CALLOC(aCount, 1);
 }
 JNIEXPORT void JNICALL Java_jse_clib_CPointer_free_1(JNIEnv *aEnv, jclass aClazz, jlong aPtr) {
     FREE((void *)(intptr_t)aPtr);
